refactor creation of lucene directory and simplify different directories implemenation strcture
