+++ modules/elasticsearch/src/main/java/org/elasticsearch/index/store/memory/ByteBufferIndexStore.java
@@ -27,7 +27,7 @@ import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.index.service.IndexService;
 import org.elasticsearch.index.settings.IndexSettings;
-import org.elasticsearch.index.store.Store;
+import org.elasticsearch.index.store.DirectoryService;
 import org.elasticsearch.index.store.support.AbstractIndexStore;
 import org.elasticsearch.monitor.jvm.JvmInfo;
 import org.elasticsearch.monitor.jvm.JvmStats;
@@ -49,8 +49,8 @@ public class ByteBufferIndexStore extends AbstractIndexStore {
         return false;
     }
 
-    @Override public Class<? extends Store> shardStoreClass() {
-        return ByteBufferStore.class;
+    @Override public Class<? extends DirectoryService> shardDirectory() {
+        return ByteBufferDirectoryService.class;
     }
 
     @Override public ByteSizeValue backingStoreTotalSpace() {
