+++ modules/elasticsearch/src/main/java/org/elasticsearch/index/store/ram/RamIndexStore.java
@@ -25,7 +25,7 @@ import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.index.Index;
 import org.elasticsearch.index.service.IndexService;
 import org.elasticsearch.index.settings.IndexSettings;
-import org.elasticsearch.index.store.Store;
+import org.elasticsearch.index.store.DirectoryService;
 import org.elasticsearch.index.store.support.AbstractIndexStore;
 import org.elasticsearch.monitor.jvm.JvmInfo;
 import org.elasticsearch.monitor.jvm.JvmStats;
@@ -43,8 +43,8 @@ public class RamIndexStore extends AbstractIndexStore {
         return false;
     }
 
-    @Override public Class<? extends Store> shardStoreClass() {
-        return RamStore.class;
+    @Override public Class<? extends DirectoryService> shardDirectory() {
+        return RamDirectoryService.class;
     }
 
     @Override public ByteSizeValue backingStoreTotalSpace() {
