+++ modules/elasticsearch/src/main/java/org/elasticsearch/indices/recovery/RecoveryTarget.java
@@ -42,7 +42,7 @@ import org.elasticsearch.index.shard.IndexShardState;
 import org.elasticsearch.index.shard.ShardId;
 import org.elasticsearch.index.shard.service.IndexShard;
 import org.elasticsearch.index.shard.service.InternalIndexShard;
-import org.elasticsearch.index.store.support.AbstractStore;
+import org.elasticsearch.index.store.Store;
 import org.elasticsearch.index.translog.Translog;
 import org.elasticsearch.indices.IndexMissingException;
 import org.elasticsearch.indices.IndicesLifecycle;
@@ -452,7 +452,7 @@ public class RecoveryTarget extends AbstractComponent {
 
             for (String existingFile : shard.store().directory().listAll()) {
                 // don't delete snapshot file, or the checksums file (note, this is extra protection since the Store won't delete checksum)
-                if (!request.snapshotFiles().contains(existingFile) && !AbstractStore.isChecksum(existingFile)) {
+                if (!request.snapshotFiles().contains(existingFile) && !Store.isChecksum(existingFile)) {
                     try {
                         shard.store().directory().deleteFile(existingFile);
                     } catch (Exception e) {
